Design Review feedback + modal refactoring
