+++ app/controllers/extraction_definitions_controller.rb
@@ -98,7 +98,7 @@ class ExtractionDefinitionsController < ApplicationController
   def extraction_definition_params
     safe_params = params.require(:extraction_definition).permit(
       :pipeline_id, :name, :format, :base_url, :throttle, :page, :per_page,
-      :total_selector, :kind, :destination_id, :source_id, :enrichment_url, :paginated, :split, :record_selector
+      :total_selector, :kind, :destination_id, :source_id, :enrichment_url, :paginated, :split, :split_selector
     )
     merge_last_edited_by(safe_params)
   end
