+++ app/frontend/entrypoints/application.js
@@ -36,7 +36,13 @@ import "/js/TestDestination";
 import "/js/Tooltips";
 import "/js/Toasts";
 import "/js/CollapseScroll";
-import "/js/formModals";
+
+import "/js/modals/updateExtractionDefinitionModal";
+import "/js/modals/createModal";
+import "/js/modals/addHarvestModal";
+import "/js/modals/transformationDefinitionSettingsModal";
+import "/js/modals/updateExtractionDefinitionModal";
+
 import "/js/SubmittingSelect";
 import "/js/AutoComplete";
