+++ app/models/extraction_definition.rb
@@ -34,7 +34,7 @@ class ExtractionDefinition < ApplicationRecord
   }.freeze
 
   validates :name, uniqueness: true
-  validates :record_selector, presence: true, if: :split?
+  validates :split_selector, presence: true, if: :split?
 
   validates :throttle, numericality: { only_integer: true, greater_than_or_equal_to: 0, less_than_or_equal_to: 60_000 }
