+++ app/views/extraction_definitions/_create_edit_harvest_modal.html.erb
@@ -98,7 +98,7 @@
                   ), {}, class: 'form-select' %>
             </div>
 
-            <div class="col-4 js-split d-none">
+            <div class="col-4 js-split d-none js-extraction-definition-split-dropdown-container">
               <%= form.label :split, class: 'form-label' do %>
                 Split
 
@@ -108,27 +108,30 @@
                 </span>
               <% end %>
             </div>
-            <div class="col-8 js-split d-none">
-              <%= form.check_box :split, class: 'form-check-input', id: 'js-extraction-definition-split-checkbox' %>
+
+            <div class="col-8 js-split d-none js-extraction-definition-split-dropdown-container">
+              <%= form.select :split, options_for_select(
+                    [['No', 'false'], ['Yes', 'true']], model.split
+                  ), {}, class: 'form-select', id: 'js-extraction-definition-split-dropdown' %>
             </div>
 
-            <div class="col-4 js-split d-none">
-              <%= form.label :record_selector, class: 'form-label' do %>
-                Record Selector
+            <div class="col-4 js-split d-none js-extraction-definition-split-selector-container">
+              <%= form.label :split_selector, class: 'form-label' do %>
+                Split Selector
 
                 <span
                     data-bs-toggle="tooltip"
-                    data-bs-title="The record selector used when splitting the file"><i class="bi bi-question-circle" aria-label="helper text"></i>
+                    data-bs-title="This selector is applied to the extracted file to determine how to split records"><i class="bi bi-question-circle" aria-label="helper text"></i>
                 </span>
               <% end %>
             </div>
 
-            <div class='col-8 js-split d-none'>
-              <%= form.text_field :record_selector,
-                                  id: 'js-extraction-definition-record-selector',
+            <div class='col-8 js-split d-none js-extraction-definition-split-selector-container'>
+              <%= form.text_field :split_selector,
+                                  id: 'js-extraction-definition-split-selector',
                                   class: {
                                     'form-control': true,
-                                    'is-invalid': @harvest_definition.errors[:record_selector].any?
+                                    'is-invalid': @harvest_definition.errors[:split_selector].any?
                                   } %>
             </div>
