+++ db/migrate/20231031013240_add_split_fields_to_extraction_definition.rb
@@ -3,6 +3,6 @@
 class AddSplitFieldsToExtractionDefinition < ActiveRecord::Migration[7.0]
   def change
     add_column :extraction_definitions, :split, :boolean, default: false, null: false
-    add_column :extraction_definitions, :record_selector, :string
+    add_column :extraction_definitions, :split_selector, :string
   end
 end
